pyproject.toml
[project]
name = "real-time-ml-system-cohort-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "quixstreams>=3.13.1",
    "websocket-client>=1.8.0",
]

[tool.uv.workspace]
members = ["services/trades"]
